backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "meridian-backend"
version = "0.1.0"
description = "FastAPI backend for Meridian"
readme = { text = "FastAPI backend for Meridian.", content-type = "text/markdown" }
requires-python = ">=3.11"
dependencies = [
  "asyncpg>=0.30,<1.0",
  "fastapi>=0.115,<1.0",
  "PyJWT[crypto]>=2.10,<3.0",
  "pydantic-settings>=2.8,<3.0",
  "sqlalchemy>=2.0.40,<3.0",
  "uvicorn[standard]>=0.34,<1.0",
]

[project.optional-dependencies]
dev = [
  "aiosqlite>=0.20,<1.0",
  "httpx>=0.28,<1.0",
  "pytest>=8.3,<9.0",
  "pytest-asyncio>=0.24,<1.0",
  "ruff>=0.11,<1.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.hatch.build.targets.wheel]
packages = ["app"]
